Refactor project and rename directories
