--- app/src/main/java/br/com/angelorobson/alternativescene/application/modules/collect/di/modules/CollectViewModelModule.kt
@@ -1,23 +1,0 @@
-package br.com.angelorobson.alternativescene.application.modules.collect.di.modules
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import br.com.angelorobson.alternativescene.application.commom.di.utils.ViewModelFactory
-import br.com.angelorobson.alternativescene.application.commom.di.utils.ViewModelKey
-import br.com.angelorobson.alternativescene.application.modules.collect.CollectViewModel
-import dagger.Binds
-import dagger.Module
-import dagger.multibindings.IntoMap
-
-@Module
-abstract class CollectViewModelModule {
-
-    @Binds
-    internal abstract fun bindViewModelFactory(factory: ViewModelFactory): ViewModelProvider.Factory
-
-    @Binds
-    @IntoMap
-    @ViewModelKey(CollectViewModel::class)
-    internal abstract fun collectViewModel(collectViewModel: CollectViewModel): ViewModel
-
-}
