--- app/src/main/java/br/com/angelorobson/alternativescene/application/modules/invoices/invoices/di/modules/InvoicesViewModelModule.kt
@@ -1,22 +1,0 @@
-package br.com.angelorobson.alternativescene.application.modules.invoices.invoices.di.modules
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import br.com.angelorobson.alternativescene.application.commom.di.utils.ViewModelFactory
-import br.com.angelorobson.alternativescene.application.commom.di.utils.ViewModelKey
-import br.com.angelorobson.alternativescene.application.modules.invoices.invoices.InvoicesViewModel
-import dagger.Binds
-import dagger.Module
-import dagger.multibindings.IntoMap
-
-@Module
-abstract class InvoicesViewModelModule {
-
-    @Binds
-    internal abstract fun bindViewModelFactory(factory: ViewModelFactory): ViewModelProvider.Factory
-
-    @Binds
-    @IntoMap
-    @ViewModelKey(InvoicesViewModel::class)
-    internal abstract fun invoicesViewModel(invoicesViewModel: InvoicesViewModel): ViewModel
-}
