+++ app/src/main/java/br/com/angelorobson/alternativescene/application/partials/events/events/di/component/EventsComponent.kt
@@ -1,0 +1,14 @@
+package br.com.angelorobson.alternativescene.application.partials.events.events.di.component
+
+import br.com.angelorobson.alternativescene.application.commom.di.modules.application.ContextModule
+import br.com.angelorobson.alternativescene.application.partials.events.events.EventsFragment
+import br.com.angelorobson.alternativescene.application.partials.events.events.di.modules.EventsModule
+import dagger.Component
+import javax.inject.Singleton
+
+@Singleton
+@Component(modules = [EventsModule::class, ContextModule::class])
+interface EventsComponent {
+
+    fun inject(eventsFragment: EventsFragment)
+}
