--- app/src/main/res/layout/account_fragment.xml
@@ -1,164 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
-        xmlns:app="http://schemas.android.com/apk/res-auto">
-    <data>
-        <variable name="user" type="br.com.angelorobson.alternativescene.domain.User"/>
-    </data>
-
-    <ScrollView android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:fillViewport="true">
-
-        <androidx.appcompat.widget.LinearLayoutCompat
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical"
-                tools:context=".application.modules.account.AccountFragment"
-                android:id="@+id/frameLayout2">
-
-            <TextView
-                    tools:text="Dados Pessoais"
-                    android:text="@string/personal_data"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_gravity="center"
-                    android:id="@+id/PersonalTextView"
-                    android:textAppearance="@style/TextAppearance.AppCompat.Large"
-                    android:textStyle="bold"
-                    android:layout_marginTop="16dp"
-                    android:layout_marginBottom="12dp"
-            />
-
-            <TextView
-                    android:id="@+id/textView2"
-                    android:text="@{user.name}"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="5dp"
-                    android:layout_gravity="center"
-                    android:gravity="center"
-                    tools:text="José da Silva"
-                    android:textColor="@color/background_color"
-                    android:textSize="18sp"
-            />
-            <TextView
-                    android:text="@{user.phoneNumber}"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="5dp"
-                    android:layout_gravity="center"
-                    android:gravity="center"
-                    tools:text="(82)95587-6548"
-                    android:textColor="@color/background_color"
-                    android:textSize="18sp"
-            />
-            <TextView
-                    android:text="@{user.email}"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="5dp"
-                    android:layout_gravity="center"
-                    android:gravity="center"
-                    tools:text="angelorobsonmelo@gmail.com"
-                    android:textColor="@color/background_color"
-                    android:textSize="18sp"
-            />
-
-            <TextView
-                    android:text="@{@string/document_number(String.valueOf(user.documentNumber))}"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="5dp"
-                    android:layout_gravity="center"
-                    android:gravity="center"
-                    tools:text="087-867-551-85"
-                    android:textColor="@color/background_color"
-                    android:textSize="18sp"
-            />
-
-            <View
-                    android:layout_width="match_parent"
-                    android:layout_height="1dp"
-                    android:layout_marginTop="15dp"
-                    android:layout_marginStart="8dp"
-                    android:layout_marginEnd="8dp"
-                    android:background="@android:color/darker_gray"/>
-
-
-            <androidx.appcompat.widget.LinearLayoutCompat
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:orientation="vertical"
-                    tools:context=".application.modules.account.AccountFragment">
-
-                <TextView
-                        tools:text="Dados Bancários"
-                        android:text="@string/bank_data"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_gravity="center"
-                        android:textAppearance="@style/TextAppearance.AppCompat.Large"
-                        android:textStyle="bold"
-                        app:layout_constraintTop_toTopOf="parent"
-                        android:layout_marginTop="16dp"
-                        android:layout_marginBottom="12dp"
-                />
-
-                <TextView
-                        android:text="@{user.bankSlug}"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_marginTop="5dp"
-                        android:layout_gravity="center"
-                        android:gravity="center"
-                        tools:text="Itaú"
-                        android:textColor="@color/background_color"
-                        android:textSize="18sp"
-                />
-
-                <TextView
-                        android:text="@{String.valueOf(user.bankAgency)}"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_marginTop="5dp"
-                        android:layout_gravity="center"
-                        android:gravity="center"
-                        tools:text="2531"
-                        android:textColor="@color/background_color"
-                        android:textSize="18sp"
-                />
-
-                <TextView
-                        android:text="@{String.valueOf(user.bankAccount)}"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_marginTop="5dp"
-                        android:layout_gravity="center"
-                        android:gravity="center"
-                        tools:text="18595400"
-                        android:textColor="@color/background_color"
-                        android:textSize="18sp"
-                />
-
-                <TextView
-                        android:text="@{@string/bank_document_number(String.valueOf(user.bankDocumentNumber))}"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_marginTop="5dp"
-                        android:layout_gravity="center"
-                        android:gravity="center"
-                        tools:text="3"
-                        android:textColor="@color/background_color"
-                        android:textSize="18sp"
-                />
-
-
-            </androidx.appcompat.widget.LinearLayoutCompat>
-
-
-        </androidx.appcompat.widget.LinearLayoutCompat>
-
-    </ScrollView>
-</layout>
