--- app/src/main/res/layout/collect_fragment.xml
@@ -1,99 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
->
-
-    <data>
-        <variable name="viewModel"
-                  type="br.com.angelorobson.alternativescene.application.modules.collect.CollectViewModel"/>
-
-        <variable name="handler"
-                  type="br.com.angelorobson.alternativescene.application.modules.collect.CollectHandler"/>
-
-        <variable name="collectRequest"
-                  type="br.com.angelorobson.alternativescene.domain.request.CollectRequest"/>
-
-    </data>
-
-    <ScrollView
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-        <androidx.appcompat.widget.LinearLayoutCompat android:layout_width="match_parent"
-                                                      android:layout_height="wrap_content"
-                                                      android:orientation="vertical"
-                                                      android:layout_margin="20dp">
-            <com.google.android.material.button.MaterialButton
-                    android:id="@+id/get_clients_button"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:text="@string/clients"
-                    android:onClick="@{() -> handler.onPressShowClientsDialog()}"
-                    style="@style/Widget.AppCompat.Button.Borderless"
-                    android:textColor="@color/white"/>
-
-
-            <com.google.android.material.textfield.TextInputLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox">
-                <ru.kolotnev.formattedittext.CurrencyEditText
-                        android:id="@+id/value_edit_text"
-                        android:layout_width="match_parent"
-                        android:layout_height="match_parent"
-                        app:validateEmpty="@{true}"
-                        app:validateTypeAutoDismiss="@{true}"
-                        android:text="@={collectRequest.valueCurrency}"
-                        android:inputType="number"
-                        android:maxLines="1"
-                        android:singleLine="true"
-                        android:hint="@string/value"
-                />
-
-            </com.google.android.material.textfield.TextInputLayout>
-
-            <com.google.android.material.textfield.TextInputLayout
-                    android:id="@+id/description_textInputLayout"
-                    android:layout_width="match_parent"
-                    android:layout_height="138dp"
-                    android:background="@drawable/rounded_border"
-                    app:layout_constraintStart_toStartOf="parent"
-                    android:paddingVertical="0dp"
-                    android:layout_marginLeft="16dp"
-                    android:layout_marginStart="16dp"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    android:layout_marginEnd="16dp"
-                    android:layout_marginRight="16dp"
-                    app:layout_constraintHorizontal_bias="0.0"
-                    android:layout_marginTop="16dp"
-                    android:hint="Observação">
-
-                <com.google.android.material.textfield.TextInputEditText
-                        android:layout_width="match_parent"
-                        android:layout_height="140dp"
-                        android:inputType="text|textMultiLine"
-                        android:gravity="top"
-                        android:overScrollMode="always"
-                        android:scrollbarStyle="insideInset"
-                        android:scrollbars="vertical"
-                        app:validateEmpty="@{true}"
-                        android:text="@={collectRequest.description}"
-                />
-
-            </com.google.android.material.textfield.TextInputLayout>
-
-
-            <com.google.android.material.button.MaterialButton
-                    android:id="@+id/save_collect_button"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:text="Cobrar"
-                    android:onClick="@{() -> handler.onPressSaveCollect(collectRequest)}"
-                    android:textColor="@color/white"/>
-
-        </androidx.appcompat.widget.LinearLayoutCompat>
-
-    </ScrollView>
-
-</layout>
