+++ app/src/main/res/layout/event_fragment.xml
@@ -15,7 +15,7 @@
         <androidx.constraintlayout.widget.ConstraintLayout
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                tools:context=".application.modules.events.event.EventFragment">
+                tools:context=".application.partials.events.event.EventFragment">
 
 
             <TextView
