--- app/src/main/res/layout/invoice_item.xml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:tools="http://schemas.android.com/tools"
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <data>
-        <variable
-                name="invoice"
-                type="br.com.angelorobson.alternativescene.domain.response.InvoiceResponse"/>
-    </data>
-
-    <TableLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="16dp"
-            android:stretchColumns="*">
-
-        <TableRow
-                android:id="@+id/tableRow1"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="10dp"
-                android:layout_marginBottom="10dp"
-                android:layout_width="0dp">
-
-            <TextView
-                    tools:text="15/03/2019"
-                    android:text="@{invoice.dateTime}"
-                    android:layout_height="wrap_content"
-                    android:layout_width="0dip"
-                    android:layout_weight="1"
-                    android:layout_gravity="center"
-                    android:textColor="#000000"
-                    android:gravity="center"/>
-
-            <TextView
-                    tools:text="Ângelo Melo"
-                    android:text="@{invoice.client.name}"
-                    android:layout_height="wrap_content"
-                    android:layout_width="0dip"
-                    android:layout_weight="1"
-                    android:layout_gravity="center"
-                    android:textColor="#000000"
-                    android:gravity="center"/>
-
-            <TextView
-
-                    android:text="@{String.valueOf(invoice.value)}"
-                    android:layout_gravity="center"
-                    android:layout_height="wrap_content"
-                    android:layout_width="0dip"
-                    android:layout_weight="1"
-                    tools:text="R$15.00,00"
-                    android:textColor="#000000"
-                     android:gravity="center"/>
-
-            <ImageView
-                    android:layout_height="wrap_content"
-                    android:layout_width="0dip"
-                    android:layout_weight="1"
-                    android:layout_gravity="center"
-                    tools:src="@drawable/ic_check"
-                    app:status="@{invoice.status}"
-                    tools:layout_editor_absoluteY="16dp"
-                    tools:layout_editor_absoluteX="371dp"/>
-
-        </TableRow>
-
-    </TableLayout>
-
-</layout>
