--- app/src/main/res/layout/invoices_fragment.xml
@@ -1,87 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-                name="viewModel"
-                type="br.com.angelorobson.alternativescene.application.modules.invoices.invoices.InvoicesViewModel"/>
-
-    </data>
-
-    <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:id="@+id/swipe_to_refresh_layout"
-            app:onRefreshListener="@{() -> viewModel.getInvoices()}"
-            app:refreshing="@{safeUnbox(viewModel.isLoadingObserver)}"
-            android:orientation="vertical">
-
-        <TableLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:stretchColumns="*">
-
-            <TableRow
-                    android:id="@+id/tableRow1"
-                    android:paddingTop="24dp"
-                    android:layout_height="wrap_content"
-                    android:layout_width="0dp">
-
-                <TextView
-                        tools:text="Data/hora"
-                        android:text="@string/date_time"
-                        android:layout_height="wrap_content"
-                        android:layout_width="0dip"
-                        android:layout_weight="1"
-                        android:layout_gravity="center"
-                        android:textColor="@color/black"
-                        android:gravity="center" android:textSize="18sp" android:textStyle="bold"/>
-
-                <TextView
-                        tools:text="Cliente"
-                        android:text="@string/client"
-                        android:layout_height="wrap_content"
-                        android:layout_width="0dip"
-                        android:layout_weight="1"
-                        android:layout_gravity="center"
-                        android:textColor="#000000"
-                        android:gravity="center" android:textSize="18sp" android:textStyle="bold"/>
-
-                <TextView
-
-                        android:text="@string/value"
-                        android:layout_gravity="center"
-                        android:layout_height="wrap_content"
-                        android:layout_width="0dip"
-                        android:layout_weight="1"
-                        tools:text="Valor"
-                        android:textColor="#000000"
-                        android:gravity="center" android:textSize="18sp" android:textStyle="bold"/>
-
-                <TextView
-                        android:text="@string/status"
-                        android:layout_gravity="center"
-                        android:layout_height="wrap_content"
-                        android:layout_width="0dip"
-                        android:layout_weight="1"
-                        tools:text="Status"
-                        android:textColor="#000000"
-                        android:gravity="center" android:textSize="18sp" android:textStyle="bold"/>
-
-            </TableRow>
-            <androidx.recyclerview.widget.RecyclerView
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:id="@+id/invoicesRecyclerView"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintBottom_toBottomOf="parent"
-                   />
-        </TableLayout>
-
-    </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
-
-</layout>
