--- app/src/main/res/layout/login_fragment.xml
@@ -1,141 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-        <import type="android.view.View"/>
-
-        <variable
-                name="authRequest"
-                type="br.com.angelorobson.alternativescene.domain.request.AuthRequest"/>
-        <variable
-                name="handler"
-                type="br.com.angelorobson.alternativescene.application.modules.login.LoginHandler"/>
-        <variable
-                name="viewModel"
-                type="br.com.angelorobson.alternativescene.application.modules.login.LoginViewModel"/>
-    </data>
-
-    <ScrollView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-                android:id="@+id/auth_constraint_layout"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                tools:context=".AuthFragment">
-
-            <com.google.android.material.textfield.TextInputLayout
-                    android:id="@+id/auth_text_input_layout_email"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginEnd="24dp"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    android:layout_marginStart="24dp"
-                    android:layout_marginTop="200dp"
-                    app:endIconMode="custom"
-                    app:layout_constraintTop_toTopOf="parent">
-
-                <com.google.android.material.textfield.TextInputEditText
-                        android:id="@+id/auth_email_edit_text"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:maxLines="1"
-                        android:singleLine="true"
-                        android:text="@={authRequest.username}"
-                        app:validateEmpty="@{true}"
-                        app:validateTypeAutoDismiss="@{true}"
-                        android:hint="@string/e_mail"/>
-
-            </com.google.android.material.textfield.TextInputLayout>
-
-            <com.google.android.material.textfield.TextInputLayout
-                    android:id="@+id/auth_password_text_input_layout"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    app:passwordToggleEnabled="true"
-                    android:layout_marginTop="26dp"
-                    app:layout_constraintTop_toBottomOf="@+id/auth_text_input_layout_email"
-                    app:layout_constraintStart_toStartOf="@+id/auth_text_input_layout_email"
-                    app:layout_constraintEnd_toEndOf="@+id/auth_text_input_layout_email">
-
-                <com.google.android.material.textfield.TextInputEditText
-                        android:id="@+id/auth_password_edit_text"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:maxLines="1"
-                        android:singleLine="true"
-                        app:validateEmpty="@{true}"
-                        app:validateTypeAutoDismiss="@{true}"
-                        android:text="@={authRequest.password}"
-                        android:hint="@string/password"
-                        android:inputType="textPassword"
-                />
-            </com.google.android.material.textfield.TextInputLayout>
-
-            <com.google.android.material.button.MaterialButton
-                    android:id="@+id/auth_submit_button"
-                    android:layout_width="0dp"
-                    android:layout_height="60dp"
-                    android:onClick="@{() -> handler.onPressLoginButton(authRequest)}"
-                    style="@style/Widget.MaterialComponents.Button.UnelevatedButton"
-                    android:layout_marginTop="16dp"
-                    app:layout_constraintTop_toBottomOf="@+id/auth_password_text_input_layout"
-                    app:layout_constraintStart_toStartOf="parent"
-                    android:layout_marginStart="37dp"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    android:layout_marginEnd="37dp"
-                    app:layout_constraintBottom_toBottomOf="parent"
-                    android:layout_marginBottom="8dp"
-                    android:text="@string/enter"
-                    android:backgroundTint="@color/tomato"
-                    android:textSize="14sp" app:layout_constraintHorizontal_bias="1.0"
-                    app:layout_constraintVertical_bias="0.02"/>
-
-            <!--<com.facebook.login.widget.LoginButton
-                    android:id="@+id/auth_facebook_button"
-                    android:layout_width="160dp"
-                    android:layout_height="22dp"
-                    android:visibility="gone"
-                    app:layout_constraintTop_toTopOf="parent" app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintEnd_toEndOf="parent"/>-->
-
-            <com.google.android.material.button.MaterialButton
-                    android:id="@+id/facebook_button_custom"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/enter_with_facebook"
-                    android:onClick="@{handler::onPressFacebookButton}"
-                    android:layout_marginTop="8dp"
-                    app:layout_constraintTop_toBottomOf="@+id/auth_submit_button"
-                    app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="8dp"
-                    app:layout_constraintEnd_toEndOf="parent" android:layout_marginEnd="8dp"
-                    app:layout_constraintBottom_toBottomOf="parent" android:layout_marginBottom="16dp"/>
-
-            <ProgressBar
-                    style="?android:attr/progressBarStyle"
-                    android:layout_width="56dp"
-                    android:layout_height="62dp"
-                    android:id="@+id/progressBar"
-                    app:visibleOrGone="@{safeUnbox(viewModel.isLoadingObserver)}"
-                    android:layout_marginTop="8dp"
-                    app:layout_constraintTop_toTopOf="parent"
-                    android:layout_marginBottom="32dp"
-                    app:layout_constraintBottom_toBottomOf="parent"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    android:layout_marginEnd="8dp"
-                    android:layout_marginRight="8dp"
-                    app:layout_constraintStart_toStartOf="parent"
-                    android:layout_marginLeft="8dp"
-                    android:layout_marginStart="8dp"
-                    android:visibility="gone"/>
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-    </ScrollView>
-
-</layout>
